requirements.txt
# Las Vegas Food Curator (LVFC) Automation System
# Dependencies for Instagram automation and content management

# Core
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
pandas>=2.0.0

# Instagram Automation
instagrapi>=0.9.9

# Video Processing
ffmpeg-python>=0.2.0
moviepy>=1.0.3

# Web Dashboard
streamlit>=1.28.0
streamlit-webrtc>=0.47.0

# Scheduling
APScheduler>=3.10.0

# Utilities
Pillow>=10.0.0
requests>=2.31.0
python-dateutil>=2.8.0

# Logging
colorlog>=6.8.0
